requirements.txt
boto3>=1.28.0
botocore>=1.31.0
streamlit>=1.25.0
python-dotenv>=1.0.0